Refactor classes out, continue Queryable
