--- src/worker.js
@@ -1,4 +1,0 @@
-onmessage = function (event) {
-    console.log('Worker: Message received from main script');
-    console.log(event);
-}
